requirements.txt
flask
waitress
selenium
python-dotenv
flask-jwt-extended